[wx] Change libGL loading
Load libGL.so.1 instead of libGL.so, this should fix things when
DL search path is broken (Ubuntu).
dlopen("libGL.so") finds Mesa driver and dlopen("libGL.so.1") finds
Nvidia or ATI. wxWidgets (added by gcc) have a dependency to libGL.so.1
so the context is created with that driver.